SoW for Puppet actuation system
This project seeks to create an intermediate control point that will allow both humans and AI to control puppet movements.  The proposal has multiple individual efforts each of which is its own project.  This meta-project will result in the ability for either AI, Human, or prerecorded movements to be translated into puppet movements, in either physical or virtual environments.